fix:  fixes for ocAskQuestionCtrl
Rename 'ownedAnswers' to 'answersForQuestion'.  Use displayingAnswers
variable and ng-if instead of JQuery.  Rename displayCount to showCount
for consistency.  Add comment to the upper limit for possible answers.  Add
check to only allow the user to display answers once since doing it again
does nothing.  Add tests for showCount().  Update 'answersForQuestion' to
make sure that the user owns the question they are trying to get answers for.

All tests pass.